% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anosimPlot.R
\name{anosimPlot}
\alias{anosimPlot}
\title{Permutation analysis of similarities between groups}
\usage{
anosimPlot(
  ano.out,
  hcol = "light blue",
  title = "Histogram of anosim permutations"
)
}
\arguments{
\item{ano.out}{an \code{anosim} object. See details and example}

\item{hcol}{color of histogram}

\item{title}{title of histogram}
}
\value{
prints a histogram of permutation tests
}
\description{
Permutation analysis of similarities between groups
}
\examples{
library(NADA) #For example data
data(Golden)

# ROS model for each group
golden.high <- with(subset(Golden,DosageGroup=="High"),NADA::ros(Blood,BloodCen))
golden.high <- data.frame(golden.high)
golden.high$DosageGroup <- "High"

golden.low <- with(subset(Golden,DosageGroup=="Low"),NADA::ros(Blood,BloodCen))
golden.low <- data.frame(golden.low)
golden.low$DosageGroup <- "Low"

golden.ros=rbind(golden.high,golden.low)

# ANOSIM analysis
library(vegan)
golden.anosim <- with(golden.ros,anosim(modeled,DosageGroup))
summary(golden.anosim)

# Plot
anosimPlot(golden.anosim)
}
\references{
Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1st ed. John Wiley and Sons, USA, N.J.

Oksanen, J., Guillaume, F., 2018. Vegan: ecological diversity. CRAN R-Project.\url{https://cran.r-project.org/web/packages/vegan/index.html}
}
\seealso{
\link[vegan:anosim]{vegan::anosim}
}
