% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atrazine.R
\docType{data}
\name{atrazine}
\alias{atrazine}
\title{Atrazine concentrations in Nebraska ground water}
\format{
An object of class \code{data.frame} with 24 rows and 4 columns.
}
\source{
Junk et al., 1980. Areal, vertical, and temporal differences in ground water chemistry: II. Journal of Environmental Quality. 9(3) 479 - 483.
}
\usage{
data(atrazine)
}
\description{
From the \code{NADA} R-Package.

Atrazine concentrations in a series of Nebraska wells before (June) and after (September) the growing season.

Objective is to determine if concentrations increase from June to September. There is one detection limit, at 0.01 ug/L. Used in Chapters 4, 5, and 9 of the Helsel (2011) book.
}
\references{
Helsel, D.R., 2011. 2011. Statistics for Censored Environmental Data Using Minitab and R. 2nd ed. John Wiley and Sons, USA, N.J.
}
\keyword{dataset}
