% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATS.R
\name{ATS}
\alias{ATS}
\title{Akritas-Theil-Sen line for censored data.}
\usage{
ATS(
  y.var,
  y.cen,
  x.var,
  x.cen = rep(0, times = length(x.var)),
  LOG = TRUE,
  retrans = FALSE,
  xlabel = NULL,
  ylabel = NULL
)
}
\arguments{
\item{y.var}{The column of y (response variable) values plus detection limits}

\item{y.cen}{The y-variable indicators, where 1 (or \code{TRUE}) indicates a detection limit in the \code{y.var} column, and \code{0} (or \code{FALSE}) indicates a detected value in \code{y.var}.}

\item{x.var}{The column of x (explanatory variable) values plus detection limits}

\item{x.cen}{The x-variable indicators, where 1 (or \code{TRUE}) indicates a detection limit in the \code{x.var} column, and \code{0} (or \code{FALSE}) indicates a detected value in \code{x.var}.}

\item{LOG}{Indicator of whether to compute the ATS line in the original y units, or for their logarithms.  The default is to use the logarithms (\code{LOG = TRUE}).  To compute in original units, specify the option \code{LOG = FALSE} (or \code{LOG = 0}).}

\item{retrans}{Indicator of whether to retransform the plot and line back to original Y-variable units.  Not needed when \code{LOG = FALSE}. \code{retrans = FALSE} & \code{LOG = TRUE} draws the plot in logY units. \code{retrans =  TRUE} & \code{LOG = TRUE} draws the plot in original Y units.}

\item{xlabel}{Custom label for the x axis of plots.  Default is x variable column name.}

\item{ylabel}{Custom label for the y axis of plots.  Default is y variable column name.}
}
\value{
Draws a scatterplot with the fitted trend-line superimposed, the equation for the Akritas-Theil-Sen line, \verb{Kendall's tau} and \code{p-value} of the test. .
}
\description{
Computes Kendall's tau and the Akritas-Theil-Sen (ATS) line for censored data.  For one x variable regression.
}
\examples{

library(NADA) #For example data

# Both y and x are censored
data(Golden)
with(Golden, ATS(Blood, BloodCen, Kidney, KidneyCen))

# x is not censored
data(HgFish)
with(HgFish, ATS(Hg, HgCen, PctWetland))
}
\references{
Akritas, M.G., Murphy, S.A., LaValley, M.P., 1995. The Theil-Sen Estimator With Doubly Censored Data and Applications to Astronomy. Journal of the American Statistical Association 90, 170–177. https://doi.org/10.2307/2291140

Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1 edition. ed. John Wiley and Sons, USA, N.J.
}
\keyword{censored}
\keyword{trend}
