% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryClust.R
\name{binaryClust}
\alias{binaryClust}
\title{Binary Cluster Matrix}
\usage{
binaryClust(dat.frame, method = "ward.D2", group = NULL, ncluster = NULL)
}
\arguments{
\item{dat.frame}{A data frame containing only the 0/1 columns.}

\item{method}{Method of forming clusters.  The default is \code{"ward.D2"}, which is appropriate for a variety of types of data.  Another appropriate option is \verb{“average”} – average distances between cluster centers.}

\item{group}{Optional grouping variable. Sites being cluster will be represented by their group name, rather than by the row number.}

\item{ncluster}{Optional number of clusters to be differentiated on the graph. Clusters are fenced off with rectagles.}
}
\value{
Prints a cluster dendrogram based on clustering method
}
\description{
Performs clustering of a matrix of 0s and 1s, ie. the censoring indicator columns for multiple variables.
}
\examples{
data(PbHeron)

# without group specified
binaryClust(PbHeron[,4:15])

# With Group argument
binaryClust(PbHeron[,4:15],group=PbHeron$DosageGroup)
}
