% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryMDS.R
\name{binaryMDS}
\alias{binaryMDS}
\title{Plot Nonmetric Multidimensional Scaling of censored data}
\usage{
binaryMDS(dat.frame, group = NULL, title = NULL, legend.pos = "bottomleft")
}
\arguments{
\item{dat.frame}{A data frame containing only the 0/1 columns.}

\item{group}{Optional grouping variable. Sites will be represented by different colored symbols for each group.}

\item{title}{Optional title for the NMDS graph.}

\item{legend.pos}{For when group is specified, the location of the legend on the graph showing the colors representing each group’s data.  Default is “bottomleft”.  Alternatives are “topright” and “centerleft”, etc.}
}
\value{
Print NMDS plot of censored data groupings.
}
\description{
Plots an NMDS of a matrix of 0s and 1s, the censoring indicator columns for multiple variables.  Use the highest censoring limit within each column.  May have different censoring levels in different columns.
}
\details{
Prior to runnning this analysis it is suggested to consult best analysis practise when performing NMDS. As a rule of thumb, an NMDS ordination with a stress value around or above 0.2 is deemed suspect and a stress value approaching 0.3 indicates that the ordination is arbitrary. Stress values equal to or below 0.1 are considered fair, while values equal to or below 0.05 indicate good fit.
}
\examples{
\dontrun{
library(NADA) #For example data
data(Golden)

# without group specified
binaryMDS(Golden[,4:15])

# With Group argument
binaryMDS(Golden[,4:15],group=Golden$DosageGroup)
}
}
\seealso{
\link[vegan:metaMDS]{vegan::metaMDS}
}
