% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarySim.R
\name{binarySim}
\alias{binarySim}
\title{Binary similarity coefficient}
\usage{
binarySim(dat.frame)
}
\arguments{
\item{dat.frame}{A data frame containing only the 0/1 columns.}
}
\value{
Returns a binary similarity matrix.
}
\description{
Computes a simple matching similarity coefficient
}
\examples{
library(NADA) #For example data
data(Golden)

binarySim(Golden$LiverCen)

}
\seealso{
\link[vegan:designdist]{vegan::designdist}
}
