% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brumbaugh.R
\docType{data}
\name{Brumbaugh}
\alias{Brumbaugh}
\title{Brumbaugh}
\format{
An object of class \code{data.frame} with 133 rows and 14 columns.
}
\source{
Brumbaugh et al., 2001, USGS Biological Science Report BSR-2001-0009.
}
\usage{
data(Brumbaugh)
}
\description{
From the \code{NADA} R-Package.

Mercury concentrations in fish across the United States.

Objective is to determine if mercury concentrations differ by watershed land use. Can concentrations be related to water and sediment characteristics of the streams?

There are three detection limits, at 0.03, 0.05, and 0.10 ug/g wet weight. Used in Chapters 10, 11 and 12 of the Helsel (2011) book.
}
\references{
Helsel, D.R., 2011. Statistics for Censored Environmental Data using Minitab and R, 2nd ed. John Wiley & Sons, USA, N.J.
}
\keyword{dataset}
