% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cadmium.R
\docType{data}
\name{Cadmium}
\alias{Cadmium}
\title{Cadmium}
\format{
An object of class \code{data.frame} with 19 rows and 3 columns.
}
\usage{
data(Cadmium)
}
\description{
From the \code{NADA} R-Package.

Cadmium concentrations in fish for two regions of the Rocky Mountains.

Objective is to determine if concentrations are the same or different in
fish livers of the two regions. There are four detection limits, at 0.2,
0.3, 0.4, and 0.6 ug/L. Used in Chapter 9 of the NADA book.
}
\references{
Helsel, D.R., 2011. Statistics for Censored Environmental Data using Minitab and R, 2nd ed. John Wiley & Sons, USA, N.J.
}
\keyword{dataset}
