% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cboxplot.R
\name{cboxplot}
\alias{cboxplot}
\title{Draws censored boxplots}
\usage{
cboxplot(
  y1,
  y2,
  group = NULL,
  LOG = FALSE,
  show = FALSE,
  ordr = NULL,
  Ylab = yname,
  Xlab = gname,
  Title = NULL,
  dl.loc = "topright",
  dl.col = "red",
  bxcol = "white",
  Ymax = NULL,
  minmax = NULL
)
}
\arguments{
\item{y1}{The column of y (response variable) values plus detection limits.}

\item{y2}{The y-variable censoring indicators, where 1 (or \code{TRUE}) indicates a detection limit in the y1 column, and 0 (or \code{FALSE}) indicates a detected value in y1.}

\item{group}{An optional column of a grouping variable.  Draws side-by-side boxplots if this variable is present.}

\item{LOG}{\code{TRUE}/\code{FALSE} indicator of whether to plot the Y axis data on the original scale (\code{FALSE}) or log scale (\code{TRUE}).}

\item{show}{\code{TRUE}\/\code{FALSE} indicator of whether to show estimated values for the portion of the box below the maximum DL (\code{TRUE}), or just leave the lower portion blank (\code{FALSE}).}

\item{ordr}{A vector indicating the order of boxes to be drawn on the boxplot, if not in alphabetical order (the default).  Example: for 4 boxplots for groups A, B, C, D, to change the order to the}

\item{Ylab}{Y axis label text, if something is wanted other than the Y variable name in the dataset.}

\item{Xlab}{X axis label text, if something is wanted other than the group variable name in the dataset.}

\item{Title}{Text to show as the graph title.  Default is blank.}

\item{dl.loc}{Location indicator of where to plot the \code{MaxDL=0.02} text on some versions of the plot.  Possible entries are “topleft”, “topright”, “topcenter”, and the corresponding “bottom” text.}

\item{dl.col}{Color of the max detection limit line(s), and the legend text stating the max DL.  Default is “red”, but it recognizes most any color by name that you can think of.}

\item{bxcol}{Color for interior of boxplots. Specify just one color if all boxes are to be the same color.  If a different color is desired for each of three boxplots, as one example, use bxcol = c(“red”, “white”, “blue”) etc.}

\item{Ymax}{Maximum Y value to be shown on the plot.  Used to cut off high outliers on plot and better show the bulk of the boxplots.}

\item{minmax}{\code{NULL}/\code{FALSE} indicator of whether to draw outliers individually. Default is to show outliers. Setting \code{minmax = FALSE} (or any text) will draw the whiskers out to the max and min of the dataset.}
}
\value{
Prints a boxplot with detection limit identified and a concatenated list of the maximum detection limit for each group.
}
\description{
Portions below the maximum detection limit(s) are not shown by default, as their percentiles are not known.
}
\details{
Note:  if one group has fewer than 3 detected observations its boxplot will not be drawn.  Its detection limits will not count when computing the maximum limit.  However, if only one boxplot is drawn for the entire dataset by not specifying a group variable, the detection limits from the portion that is the mostly ND group will be used when computing the maximum limit.
}
\examples{

library(NADA) #For example data
data(Golden)
cboxplot(Golden$Liver,Golden$LiverCen,Golden$Group)
}
