% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cen1way.R
\name{cen1way}
\alias{cen1way}
\title{Peto-Peto nonparametric test}
\usage{
cen1way(y1, y2, grp)
}
\arguments{
\item{y1}{The column of data values plus detection limits}

\item{y2}{The column of indicators, where 1 (or \code{TRUE}) indicates a detection limit in the y1 column, and 0 (or \code{FALSE}) indicates a detected value in y1.}

\item{grp}{Grouping or factor variable. Can be either a text or numeric value indicating the group assignment.}
}
\value{
1-way test to test differences between groups
}
\description{
Performs a Peto-Peto nonparametric test of differences in cdfs between groups.  If more than two groups, the test is followed by a nonparametric multiple comparison test.  Uses the BH method of adjusting p-values.
}
\examples{

library(NADA) #for example data
data(Golden)

# Two Groups
cen1way(Golden$Liver,Golden$LiverCen,Golden$DosageGroup)

# More than two groups
cen1way(Golden$Liver,Golden$LiverCen,Golden$Group)
}
\keyword{difference}
