% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cen2means.R
\name{cen2means}
\alias{cen2means}
\title{Censored data t-test}
\usage{
cen2means(y1, y2, grp, LOG = TRUE)
}
\arguments{
\item{y1}{The column of data values plus detection limits}

\item{y2}{The column of indicators, where 1 (or \code{TRUE}) indicates a detection limit in the y1 column, and 0 (or \code{FALSE}) indicates a detected value in y1.}

\item{grp}{Grouping or factor variable. Can be either a text or numeric value indicating the group assignment.}

\item{LOG}{Indicator of whether to compute tests in the original units, or on their logarithms.  The default is to use the logarithms (LOG = \code{TRUE}).  To compute in original units, specify the option LOG = \code{FALSE} (or LOG = 0).}
}
\value{
Censored data analogue of the t-test
}
\description{
Performs a parametric test of differences in means between two groups of censored data, either in original or in log units (the latter becomes a test for difference in geometric means).
}
\details{
Note that because this is an MLE procedure, when a normal distribution model is used (LOG=FALSE) values may be modeled as below zero.  When this happens the p-values may be unreal (often lower than they should be).  Because of this, testing in log units is preferable and the default.
}
\examples{

library(NADA) #for example data
data(Golden)
cen2means(Golden$Liver,Golden$LiverCen,Golden$DosageGroup)
}
\keyword{t-test}
