% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cen_paired.R
\name{cen_paired}
\alias{cen_paired}
\title{Censored data paired t-test}
\usage{
cen_paired(xd, xc, yd, yc, alternative = "two.sided")
}
\arguments{
\item{xd}{The first column of data values plus detection limits}

\item{xc}{The column of censoring indicators, where 1 (or \code{TRUE}) indicates a detection limit in the xd column, and 0 (or \code{FALSE}) indicates a detected value in xd.}

\item{yd}{The second column of data values plus detection limits, or a single number representing a standard / guideline value.}

\item{yc}{The column of censoring indicators for yd, where 1 (or \code{TRUE}) indicates a detection limit in the yd column, and 0 (or \code{FALSE}) indicates a detected value in \code{yd}. Not needed if \code{yd} is a single standard number.}

\item{alternative}{The usual notation for the alternate hypothesis.  Default is \verb{“two.sided”}.  Options are \verb{“greater”} or \verb{“less”}.}
}
\value{
A list of statistics containing the following components:
\itemize{
\item \code{n} Number of samples
\item \code{Z} The value of the test statistic
\item \code{p.value} the p-value of the test
\item \verb{Mean difference} the mean difference between \code{xd} and \code{yd}
}
}
\description{
Performs a parametric test of whether the mean difference between two columns of paired censored data equals 0.
}
\details{
You may also test for whether the x data exceed a standard by entering the single number for the standard as \code{yd}.  In that case no \code{yc} is required.
}
\examples{

library(NADA) #for example data
data(Atra)

cen_paired(Atra$June,Atra$JuneCen,Atra$Sept,Atra$SeptCen)

# Comparing standard/guieline value
cen_paired(Atra$June,Atra$JuneCen,0.01)
}
\references{
Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1 edition. ed. John Wiley and Sons, USA, N.J.
}
\keyword{paired}
\keyword{t-test}
