% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cen_signedranktest.R
\name{cen_signedranktest}
\alias{cen_signedranktest}
\title{Wilcoxcon Signed-Rank test for censored data}
\usage{
cen_signedranktest(xd, xc, yd, yc, alternative = "two.sided")
}
\arguments{
\item{xd}{The first column of data values plus detection limits}

\item{xc}{The column of censoring indicators, where 1 (or \code{TRUE}) indicates a detection limit in the xd column, and 0 (or \code{FALSE}) indicates a detected value in xd.}

\item{yd}{The second column of data values plus detection limits, or a single number representing a standard / guideline value.}

\item{yc}{The column of censoring indicators for yd, where 1 (or \code{TRUE}) indicates a detection limit in the yd column, and 0 (or \code{FALSE}) indicates a detected value in \code{yd}. Not needed if \code{yd} is a single standard number.}

\item{alternative}{The usual notation for the alternate hypothesis.  Default is \verb{“two.sided”}.  Options are \verb{“greater”} or \verb{“less”}.}
}
\value{
A list of Wilcoxon Signed-Rank test with Pratt correction for ites statistics containing the following components:
\itemize{
\item \code{n} Number of samples
\item \code{Z} The value of the test statistic
\item \code{p.value} the p-value of the test
}
}
\description{
Performs a nonparametric Wilcoxon signed-rank test of whether the median difference between two columns of paired censored data equals 0.  Uses the Pratt adjustment for pairs of equal values.
}
\examples{

library(NADA) #For example data
data(Atra)

cen_signedranktest(Atra$June,Atra$JuneCen,Atra$Sept,Atra$SeptCen)
}
\keyword{Wilcoxon}
