% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cenanova.R
\name{cenanova}
\alias{cenanova}
\title{ANOVA for censored data}
\usage{
cenanova(y1, y2, grp, LOG = TRUE)
}
\arguments{
\item{y1}{The column of data values plus detection limits}

\item{y2}{The column of indicators, where 1 (or \code{TRUE}) indicates a detection limit in the \code{y1} column, and 0 (or \code{FALSE}) indicates a detected value in \code{y1}.}

\item{grp}{Grouping or factor variable. Can be either a text or numeric value indicating the group assignment.}

\item{LOG}{Indicator of whether to compute tests in the original units, or on their logarithms.  The default is to use the logarithms (\code{LOG = TRUE}).  To compute in original units, specify the option \code{LOG = FALSE} (or \code{LOG = 0}).}
}
\value{
Returns the Maximum Likelihood Estimation (MLE) comparison results including Chi-Squared value, degrees of freedom and \code{p-value} of the test. Test assumes lognormal(\code{LOG=TRUE}) or nomal(\code{LOG=FALSE}) distribution

In addition to comparison of censored data between groups, multiple comparisons of means are also printed based on the parametric survival model (\code{CensData~Factor}).
\itemize{
\item Group Names of groups (NOTE: \code{==} indicates \code{"two.sided"} comparisons\verb{) \\item }Estimate\verb{Slope of the parametric survival model \\item}Std. Error\verb{Standard error of estimate \\item}z value\verb{z value test statistic \\time}Pr(>|z|)` P-values of test
}
}
\description{
Performs a parametric test of differences in means between groups of censored data, followed by a parametric Tukey's multiple comparison test.
}
\details{
When this happens the p-values may be unreal (often lower than they should be).  Because of this, testing in log units is preferable and the default.
}
\examples{
data(PbHeron)
cenanova(PbHeron$Liver,PbHeron$LiverCen,PbHeron$DosageGroup)

cenanova(PbHeron$Liver,PbHeron$LiverCen,PbHeron$DosageGroup,LOG=FALSE)

}
\seealso{
\link[survival:survreg]{survival::survreg}
}
