% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cenCompareCdfs.R
\name{cenCompareCdfs}
\alias{cenCompareCdfs}
\title{Comparison of empirical cdf of censored data}
\usage{
cenCompareCdfs(y.var, cen.var, dist3 = "normal", Yname = yname)
}
\arguments{
\item{y.var}{The column of y (response variable) values plus detection limits}

\item{cen.var}{The column of indicators, where 1 (or \code{TRUE}) indicates a detection limit in the \code{y.var} column, and 0 (or \code{FALSE}) indicates a detected value in \code{y.var}.}

\item{dist3}{Name of the third distribution to be plotted, default is \code{normal} distrubtion. Other distributions include \code{lnorm}(for log-normal), \code{gamma} and \code{weibull}(for Weibull).}

\item{Yname}{Optional – input text in quotes to be used as the variable name.  The default is the name of the \code{y.var} input variable.}
}
\value{
prints a plot of the empirial CDFs with BIC value for each distribution.
}
\description{
Plots the empirical cdf and cdfs of three theoretical distributions, fit by maximum likelihood estimation (MLE).
}
\examples{

library(NADA) #For example data

data(TCEReg)
cenCompareCdfs(TCEReg$TCEConc, TCEReg$TCECen)

# With log-normal distribution
cenCompareCdfs(TCEReg$TCEConc, TCEReg$TCECen,dist3="lnorm")

# Using an distribution not supported by this function (yet)
cenCompareCdfs(TCEReg$TCEConc, TCEReg$TCECen,dist3="beta")

# With Yname specified
cenCompareCdfs(TCEReg$TCEConc, TCEReg$TCECen,Yname="TCE Conc (ug/L)\nLong Island, NY USA")
}
\keyword{CDF}
