% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cenCompareQQ.R
\name{cenCompareQQ}
\alias{cenCompareQQ}
\title{Censored Q-Q Plot comparison}
\usage{
cenCompareQQ(y.var, cen.var, Yname = yname)
}
\arguments{
\item{y.var}{The column of y (response variable) values plus detection limits}

\item{cen.var}{The column of indicators, where 1 (or \code{TRUE}) indicates a detection limit in the \code{y.var} column, and 0 (or \code{FALSE}) indicates a detected value in \code{y.var}.}

\item{Yname}{Optional – input text in quotes to be used as the variable name on all plots.  The default is the name of the \code{y.var} input variable.}
}
\value{
Plots three Q-Q plots based on normal, lognormal and gamma distributions and prints the best-fit distribution.
}
\description{
Produces three quantile-quantile (Q-Q) plot, also called a probability plot based on three distribution (normal, lognormal and gamma distributions).
}
\details{
Produces three Q-Q plots and reports which has the highest Shapiro-Francia test statistic (W).  The distribution with the highest W is the best fit of the three.
}
\examples{
data(Brumbaugh)
cenCompareQQ(Brumbaugh$Hg,Brumbaugh$HgCen)
}
\references{
Helsel, D.R., 2011. Statistics for censored environmental data using Minitab and R, 2nd ed. John Wiley & Sons, USA, N.J.

Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1st ed. John Wiley and Sons, USA, N.J.

Shapiro, S.S., Francia, R.S., 1972. An approximate analysis of variance test for normality. Journal of the American Statistical Association 67, 215–216.
}
