% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cencorreg.R
\name{cencorreg}
\alias{cencorreg}
\title{Correlation and Regression with censored data}
\usage{
cencorreg(y.var, cen.var, x.vars, LOG = TRUE)
}
\arguments{
\item{y.var}{The column of y (response variable) values plus detection limits.}

\item{cen.var}{The column of indicators, where 1 (or \code{TRUE}) indicates a detection limit in the \code{y.var} column, and 0 (or \code{FALSE}) indicates a detected value is in \code{y.var}.}

\item{x.vars}{One or more uncensored explanatory variable(s). See Details}

\item{LOG}{Indicator of whether to compute the regression in the original y units, or on their logarithms.  The default is to use the logarithms (\code{LOG = TRUE}).  To compute in original units, specify the option \code{LOG = FALSE} (or \code{LOG = 0}).}
}
\value{
When \code{x.vars} is more than one variable, likelihood, rescaled likelihood and McFaddens correlation coefficient (\code{R}) is printed.

When \code{x.vars} is one variable likelihood, rescaled likelihood and McFaddens coefficent of determination (\code{R2}) is printed

In addition to goodness-of-fit coefficients, model AIC and BIC values are provided.

Parametric Survival Model output is also provided including regression model.
}
\description{
Computes three parametric correlation coefficients for one X variable and the corresponding R2 for multiple X variables, and a regression equation for censored data.
}
\details{
\code{x.vars}: If 1 x variable only, enter its name.  If multiple x variables, enter the name of a data frame of columns of the x variables. No extra columns unused in the regression allowed. Create this by \code{x.frame <- data.frame (Temp, Flow, Time)} for 3 variables (temperature, flow and time).

AIC and BIC are printed to help evaluate the ‘best’ regression model.

The default is that the Y variable will be log transformed.
}
\examples{
library(NADA) #For example data

data(HgFish)

# One variable
cencorreg(HgFish$Hg,HgFish$HgCen,HgFish$SedMeHg)

# More than one variable for demostration purposes
cencorreg(HgFish$Hg,HgFish$HgCen,HgFish[,c("SedMeHg","PctWetland")])
}
\references{
Helsel, D.R., 2011. Statistics for censored environmental data using Minitab and R, 2nd ed. John Wiley & Sons, USA, N.J.

Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1st ed. John Wiley and Sons, USA, N.J.
}
\seealso{
\link[survival:survreg]{survival::survreg}
}
