% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NADA_All.R
\name{cenpar.usr}
\alias{cenpar.usr}
\title{Internal: Get Plotting Limits in Log or Linear Scale}
\usage{
cenpar.usr(log)
}
\arguments{
\item{log}{Character specifying log scale: "x", "y", or "xy".}
}
\value{
Numeric vector of plot limits adjusted for log scale.
}
\description{
Internal: Get Plotting Limits in Log or Linear Scale
}
\keyword{internal}
