% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cenPredInt.R
\name{cenPredInt}
\alias{cenPredInt}
\title{Prediction interval for censored data}
\usage{
cenPredInt(
  y.var,
  cen.var,
  pi.type = "two-sided",
  conf = 0.95,
  newobs = 1,
  method = "mle"
)
}
\arguments{
\item{y.var}{The column of y (response variable) values plus detection limits}

\item{cen.var}{The column of indicators, where 1 (or \code{TRUE}) indicates a detectionlimit in the \code{y.var} column, and 0 (or \code{FALSE}) indicates a detected value in \code{y.var}.}

\item{pi.type}{Designation of either a \verb{“two-sided”} interval (default) or a 1-sided \verb{“upper”} or 1-sided \verb{“lower”} interval.}

\item{conf}{Confidence coefficient of the interval, 0.95 (default).}

\item{newobs}{The number of new observations to be contained in the interval.}

\item{method}{character string specifying the method of estimation. Default is \code{mle} (maximum likelihood).}
}
\value{
A table of prediction limits based on user provided confidence coefficient (\code{conf}) and prediction invterval type (\code{pi.type})
}
\description{
Computes prediction intervals for censored data assuming lognormal, gamma and normal distributions.
}
\examples{

data(PbHeron)

# Default
cenPredInt(PbHeron$Liver,PbHeron$LiverCen)

# User defined confidence coefficient
cenPredInt(PbHeron$Liver,PbHeron$LiverCen, conf=0.5)

# User defined confidence coefficient outside of acceptable range
# the procedure will stop and give an error.
\dontrun{cenPredInt(PbHeron$Liver,PbHeron$LiverCen, conf=1.1)}

# User defined prediction interval type
cenPredInt(PbHeron$Liver,PbHeron$LiverCen,pi.type="lower")
cenPredInt(PbHeron$Liver,PbHeron$LiverCen,pi.type="upper")
}
\references{
Helsel, D.R., 2011. Statistics for censored environmental data using Minitab and R, 2nd ed. John Wiley & Sons, USA, N.J.

Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1st ed. John Wiley and Sons, USA, N.J.
}
\seealso{
\link[EnvStats:enormCensored]{EnvStats::enormCensored}
}
\keyword{interval}
\keyword{prediction}
