% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censeaken.R
\name{censeaken}
\alias{censeaken}
\title{Seasonal Kendall permutation test on censored data}
\usage{
censeaken(
  time,
  y,
  y.cen,
  group,
  LOG = FALSE,
  R = 4999,
  nmin = 4,
  seaplots = FALSE
)
}
\arguments{
\item{time}{Column of the time variable, either a sequence of days or decimal times, etc.  Will be the scale used for time in the trend analysis.}

\item{y}{The column of y (response variable) values plus detection limits}

\item{y.cen}{The y-variable indicators, where 1 (or \code{TRUE}) indicates a detection limit in the \code{y.var} column, and 0 (or \code{FALSE}) indicates a detected value in \code{y.var}.}

\item{group}{Column of the season classifications. A factor in R, so usually though not necessarily a text variable.  If numeric, define as a factor before running the script.}

\item{LOG}{Indicator of whether to compute the regression in the original y units, or on their logarithms.  The default is to use the logarithms (LOG = \code{TRUE}).  To compute in original units, specify the option LOG = \code{FALSE} (or LOG = 0).}

\item{R}{The number of repetitions in the permutation process.  R is often between 999 and 9999 (+ the 1 observed test statistic produces 1000 to 10000 repetitions). By default R=4999. Increasing R simply results in lower variation in the pvalues produced between runs.}

\item{nmin}{The minimum number of observations needed for the entire time period to be tested, per season.  For example, with 1 sample per year per season over an 8-year period, you have 8 observations for each season.  You can increase this number if you want a higher minimum.  Don’t decrease it below 4.  If there are fewer than nmin values that season is skipped and not included in the overall test & a note will be printed.}

\item{seaplots}{In addition to the plot of the overall Seasonal Kendall line, plots for the individual seasons can be drawn.}
}
\value{
Prints Kendall Trend test results for each season (individually) and a Seasonal Kendall test and Theil-Sen line results.

If \code{seaplots=TRUE} each season's trend line will be plotted along with overall Seasonal Kendall and Theil-Sen line.

If \code{seaplots=FALSE} the entire dataset will be plotted with Thiel-Sen line overlayed.
}
\description{
Seasonal Kendall permutation test on censored data
}
\examples{

data(Brumbaugh)

#Artifical time and season variables for demonstration purposes
Brumbaugh$time=1:nrow(Brumbaugh)
Brumbaugh$sea=as.factor(round(runif(nrow(Brumbaugh),1,4),0))

with(Brumbaugh,censeaken(time,Hg,HgCen,sea,seaplots = TRUE))
}
\references{
Helsel, D.R., 2011. Statistics for censored environmental data using Minitab and R, 2nd ed. John Wiley & Sons, USA, N.J.

Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1st ed. John Wiley and Sons, USA, N.J.
}
\seealso{
\link[NADA:cenken]{NADA::cenken}
}
