% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cenTolInt.R
\name{cenTolInt}
\alias{cenTolInt}
\title{Upper Tolerance interval for censored data}
\usage{
cenTolInt(y.var, cen.var, conf = 0.95, cover = 0.9, method.fit = "mle")
}
\arguments{
\item{y.var}{The column of y (response variable) values plus detection limits}

\item{cen.var}{The column of indicators, where 1 (or \code{TRUE}) indicates a detectionlimit in the \code{y.var} column, and 0 (or \code{FALSE}) indicates a detected value in \code{y.var}.}

\item{conf}{Confidence coefficient of the interval, 0.95 (default).}

\item{cover}{Coverage, the percentile probability above which the tolerance interval is computed.  The default is 90, so a tolerance interval will be computed above the 90th percentile of the data.}

\item{method.fit}{The method used to compute the parameters of the distribution.  The default is maximum likelihood (\verb{“mle”}). The alternative is robust ROS (\verb{“rROS”}).}
}
\value{
Plot of Empirical and theoretical CDFs with BIC values provided

Prints the percentile (\code{cover}) and upper tolerance limit (\code{conf}) for lognormal, normal and approximated gamma distributions.
}
\description{
Computes a one-sided upper tolerance interval for censored data assuming lognormal, gamma and normal distributions.
}
\examples{

data(PbHeron)

# Default
cenTolInt(PbHeron$Liver,PbHeron$LiverCen)

# User defined conficence interval
cenTolInt(PbHeron$Liver,PbHeron$LiverCen,conf=0.75)

# User defined percentile
cenTolInt(PbHeron$Liver,PbHeron$LiverCen,cover=0.5)

# inputs outside acceptable ranges
\dontrun{
cenTolInt(PbHeron$Liver,PbHeron$LiverCen,cover=1.25)
cenTolInt(PbHeron$Liver,PbHeron$LiverCen,conf=1.1)
cenTolInt(PbHeron$Liver,PbHeron$LiverCen,method.fit="ROS")
}
}
\references{
Helsel, D.R., 2011. Statistics for censored environmental data using Minitab and R, 2nd ed. John Wiley & Sons, USA, N.J.

Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1st ed. John Wiley and Sons, USA, N.J.
}
