% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfit.R
\name{cfit}
\alias{cfit}
\title{Compute an ECDF and Distribution Parameters for Censored Data}
\usage{
cfit(
  y1,
  y2,
  conf = 0.95,
  qtls = c(0.1, 0.25, 0.5, 0.75, 0.9),
  Cdf = TRUE,
  printstats = TRUE,
  Ylab = NULL
)
}
\arguments{
\item{y1}{Either the lowest possible concentrations (interval-censored format) or conccentrations plus detection limits for indicator formated data.}

\item{y2}{Either the highest possible concentrations (interval-censored format) or censoring indicators (logical. 1 or \code{TRUE} = censored, 0 or FALSE = detected) for indicator formatted data.}

\item{conf}{The confidence coefficient for confidence intervals around the Kaplan-Meier mean and median. Default = 0.95.}

\item{qtls}{Probabilities for the quantiles to be estimated.  Defaults are (0.10, 0.25, 0.50, 0.75, 0.90).  You may add and/or substitute probabilities -- all must be between and not including 0 to 1..}

\item{Cdf}{Logical \code{TRUE}/\code{FALSE} indicator of whether to plot the empirical cumulative distribution function (cdf) using Kaplan-Meier quantiles.}

\item{printstats}{Logical \code{TRUE}/\code{FALSE} option of whether to print the resulting statisics in the console window, or not.  Default is TRUE.}

\item{Ylab}{Optional input text in quotes to be used as the variable name on the ecdf plot.  The default is the name of the \code{y1} input variable.}
}
\value{
If \code{printstats=TRUE}: Based on the provided \code{conf} value, Kaplan-Meier summary statistics (\code{mean},\code{sd},\code{median}), lower and upper confidence intervals around the mean and median value, sample size and percent of censored samples are returned. The specified quantile values are also printed and returned.

If \code{Cdf=TRUE}: The ecdf of censored data is plotted.
}
\description{
Computes the empirical cumulative distribution function (ECDF) for censored data. Estimates parameters of the distribution, including the mean and quantiles.
}
\details{
Quantiles and parameters are estimated using the survfit function. The mean computed is the "restricted mean" (see help for the survfit function of the survival package).  Internally uses interval-censoring to avoid a small bias in the mean produced by the NADA package's cenfit function, which uses the reverse Kaplan-Meier procedure, converting left-censored to right-censored data prior to computing the ecdf and mean. See Gillespie et al. for more discussion.
}
\examples{

data(Brumbaugh)

cfit(Brumbaugh$Hg,Brumbaugh$HgCen)

}
\references{
Helsel, D.R., 2011. Statistics for Censored Environmental Data using Minitab and R, 2nd ed. John Wiley & Sons, USA, N.J.

Gillespie, B.W., et al., 2010.  Estimating Population Distributions When Some Data Are Below a Limit of Detection by Using a Reverse Kaplan-Meier Estimator. Epidemiology 21, 564-570.
}
\seealso{
\link[survival:survfit]{survival::survfit} \link[NADA:cenfit]{NADA::cenfit}
}
