% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NADA_ros.R
\name{cohn}
\alias{cohn}
\title{Calculate Cohn Numbers}
\usage{
cohn(obs, censored, na.action = getOption("na.action"))
}
\arguments{
\item{obs}{A numeric vector of observations.}

\item{censored}{A logical vector indicating which observations are censored.}

\item{na.action}{Function to handle missing values (default is getOption("na.action")).}
}
\value{
An object of class \code{"cohn"} containing the Cohn statistics.
}
\description{
Computes the A, B, C, P values used in probability plotting for censored data.
}
\examples{

obs <- c(1, 2, 3, 4, 5, 6, 7)
cens <- c(TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE)
NADA2:::cohn(obs, cens)

}
\keyword{internal}
