% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeS.R
\name{computeS}
\alias{computeS}
\title{Computes Kendall rank correlation S-value (score) for censored data}
\usage{
computeS(x, y, ycen, seas = NULL, R = R)
}
\arguments{
\item{x}{Column of the time variable, either a sequence of days or decimal times, etc.  Will be the scale used for time in the trend analysis.}

\item{y}{The column of y (response variable) values plus detection limits}

\item{ycen}{The y-variable indicators, where 1 (or \code{TRUE}) indicates a detection limit in the \code{y} column, and 0 (or \code{FALSE}) indicates a detected value in \code{y}.}

\item{seas}{Column of the season classifications. A factor in R, so usually though not necessarily a text variable.  If numeric, define as a factor before running the script.}

\item{R}{The number of repetitions in the permutation process.  R is often between 999 and 9999 (+ the 1 observed test statistic produces 1000 to 10000 repetitions).}
}
\value{
S-value for each permutation
}
\description{
A permutation test to compute Kendall rank correlation S-values on censored data.
}
\examples{
library(NADA) #For example data

data(HgFish)

#Artifical time and season variables for demonstration purposes
HgFish$time=1:nrow(HgFish)
HgFish$sea=as.factor(round(runif(nrow(HgFish),1,4),0))


with(HgFish,computeS(time,Hg,HgCen,sea,R=100))

}
\seealso{
\link[Kendall:Kendall]{Kendall::Kendall}
}
