% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CuZn.R
\docType{data}
\name{CuZn}
\alias{CuZn}
\title{CuZn}
\format{
An object of class \code{data.frame} with 118 rows and 5 columns.
}
\source{
Millard and Deverel, 1988, Water Resources Research 24, pp.2087-2098.
}
\usage{
data(CuZn)
}
\description{
Originally in the \code{NADA} R-Package.

Copper and zinc concentrations in ground waters from two zones in the San Joaquin Valley of California.  The zinc concentrations were used.

Objective is to determine if zinc concentrations differ between the two zones.  Zinc has two detection limits, at 3 and 10 ug/L.  Used in Chapters 4, 5 and 9 of the NADA book.
}
\references{
Helsel, D.R., 2005.Nondectects and Data Analysis; Statistics for censored environmental data. John Wiley & Sons, USA, N.J.
}
\keyword{dataset}
