% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalent_n.R
\name{equivalent_n}
\alias{equivalent_n}
\title{Censored data sample size}
\usage{
equivalent_n(y.var, y.cen)
}
\arguments{
\item{y.var}{The column of data values plus detection limits.}

\item{y.cen}{The column of indicators, where 1 (or \code{TRUE}) indicates a detection limit in the \code{y.var} column, and 0 (or \code{FALSE}) indicates a detected value in \code{y.var}.}
}
\value{
Prints summary statistics including
\itemize{
\item \code{n} sample size
\item \code{n.cen} number of censored data
\item \code{pct.cen} precent censored data
\item \code{min} minimum reported value
\item \code{max} maximum reported value
}

Summary of censored data including
\itemize{
\item \code{limit} detection limit
\item \code{n} number of censored values per limit
\item \code{uncen} number of uncensored values
\item \code{pexceed} proportion that exceeds the limit
}

Summary of the equivalent sample size for detected and censored values.
}
\description{
Computes the equivalent sample size of censored data.  Observations at lower detection limits have more information than observations at higher detection limits.
}
\details{
Based on "Method 2" of Dr. Brenda Gillespie's talk at ASA National Meeting 2019.

Differs in the method for computing the percentile probability for the detection limits.

Computes the equivalent n, the number of observations including censored values, as a measure of information content for data with nondetects.
}
\examples{
data(Brumbaugh)

equivalent_n(Brumbaugh$Hg,Brumbaugh$HgCen)
}
\references{
Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1st ed. John Wiley and Sons, USA, N.J.
}
\seealso{
\link[NADA:censummary]{NADA::censummary}
}
\keyword{Sample}
\keyword{Size}
\keyword{censored}
