% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example1.R
\docType{data}
\name{Example1}
\alias{Example1}
\title{Example1}
\format{
An object of class \code{data.frame} with 21 rows and 3 columns.
}
\source{
unknown
}
\usage{
data(Example1)
}
\description{
Arsenic concentrations (possibly artificial) in groundwater.

Objective -- To test whether the mean concentration can be show to have exceeded the
10 microgram per liter health standard.
}
\references{
unknown
}
\keyword{dataset}
