% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example2.R
\docType{data}
\name{Example2}
\alias{Example2}
\title{Example2}
\format{
An object of class \code{data.frame} with 31 rows and 5 columns.
}
\source{
unknown
}
\usage{
data(Example2)
}
\description{
Methyl Isobutyl Ketone (MIBK) was measured in air above a medium-sized US city. Data were
reported only as "ND" or as a detected concentration (no information on reporting limits).

Objective -- To compute an upper confidence limit on the mean without having the value(s)
of the reporting limit(s).
}
\references{
unknown
}
\keyword{dataset}
