% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example3.R
\docType{data}
\name{Example3}
\alias{Example3}
\title{Example3}
\format{
An object of class \code{data.frame} with 14 rows and 4 columns.
}
\source{
unknown
}
\usage{
data(Example3)
}
\description{
Arsenic concentrations in ground water from a private supply well.  All 14 observations
are below one of several reporting limits (1005 non-detects).

Objective -- To show what type of comparison to a health standard can be made using a
dataset without detected observations.
}
\references{
unknown
}
\keyword{dataset}
