% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gales_Creek.R
\docType{data}
\name{Gales_Creek}
\alias{Gales_Creek}
\title{Gales_Creek}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 63 rows and 11 columns.
}
\source{
US Geological Survey.  https://waterdata.usgs.gov/monitoring-location/453026123063401/
}
\usage{
data(Gales_Creek)
}
\description{
Total recoverable chromium concentrations in streamflows of Gales Creek in Oregon, USA.

Objective is to relate chromium concentrations including censored values to mean daily flows over time and by season (wet versus dry seasons).  Two detection limits decreasing over time.
}
\references{
Publicly available data.
}
\keyword{dataset}
