% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NADA_ros.R
\name{hc_ppoints}
\alias{hc_ppoints}
\title{Compute plotting positions for censored and uncensored data}
\usage{
hc_ppoints(obs, censored, na.action = getOption("na.action"))
}
\arguments{
\item{obs}{A numeric vector of observations.}

\item{censored}{A logical vector indicating which values are censored.}

\item{na.action}{A function to handle NA values (default from options).}
}
\value{
A numeric vector of plotting positions.
}
\description{
Compute plotting positions for censored and uncensored data
}
\keyword{internal}
