% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenplot.R
\name{kenplot}
\alias{kenplot}
\title{Plot sensored trend for censored data}
\usage{
kenplot(y1, ycen, x1, xcen, retrans = FALSE, xnam = NULL, ynam = NULL)
}
\arguments{
\item{y1}{The column of y (response variable) values plus detection limits}

\item{ycen}{The y-variable indicators, where 1 (or \code{TRUE}) indicates a detection limit in the \code{y.var} column, and \code{0} (or \code{FALSE}) indicates a detected value in \code{y.var}.}

\item{x1}{The column of x (explanatory variable) values plus detection limits}

\item{xcen}{The x-variable indicators, where 1 (or \code{TRUE}) indicates a detection limit in the \code{x.var} column, and \code{0} (or \code{FALSE}) indicates a detected value in \code{x.var}.}

\item{retrans}{Indicator of whether to retransform the plot and line back to original Y-variable units.  Not needed when \code{LOG = FALSE}. \code{retrans = FALSE} & \code{LOG = TRUE} draws the plot in logY units. \code{retrans =  TRUE} & \code{LOG = TRUE} draws the plot in original Y units.}

\item{xnam}{Custom label for the x axis of plots.  Default is x variable column name.}

\item{ynam}{Custom label for the y axis of plots.  Default is y variable column name.}
}
\value{
#' @references
Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1st ed. John Wiley and Sons, USA, N.J.
}
\description{
Computes Kendall's tau and the Akritas-Theil-Sen (ATS) line for censored data.  For one x variable regression.
}
\examples{
\dontrun{
# Both y and x are censored
data(PbHeron)
with(PbHeron, kenplot(Blood, BloodCen, Kidney, KidneyCen))

# x is not censored
data(Brumbaugh)
with(Brumbaugh, kenplot(Hg, HgCen, PctWetland,rep(0, times=length(PctWetland))))
}
}
\seealso{
\link[NADA:cenken]{NADA::cenken}
}
