% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NADA_icmle.R
\name{Macmat}
\alias{Macmat}
\title{Create clique matrix and Petrie pairs from maximal antichains}
\usage{
Macmat(ml)
}
\arguments{
\item{ml}{A list of maximal antichains as returned by \code{\link[=Maclist]{Maclist()}}.}
}
\value{
A list of class \code{petrie} containing:
\describe{
\item{pmat}{Clique matrix}
\item{ppairs}{Start and end of each interval in antichains}
}
}
\description{
Create clique matrix and Petrie pairs from maximal antichains
}
\keyword{internal}
