% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markers.R
\docType{data}
\name{markers}
\alias{markers}
\alias{Markers}
\title{Markers}
\format{
An object of class \code{data.frame} with 30 rows and 15 columns.
}
\source{
Symonds et al. (2016) Journ Applied Microbio  121, p. 1469-1481.
}
\usage{
data(Markers)
}
\description{
Three MST markers of human-origin contamination, and three of animal contamination, were
evaluated along with a general fecal indicator to determine if the pattern of their
levels differed among five coastal sites off the coast of Florida. Data are interval-censored.

Objective -- Test whether the pattern of six microbial source tracking (MST) markers
is related to the Entero1A indicator of general fecal pollution.  Also determine if the
pattern of markers plus Entero1A indicator differs among five coastal water sites using ANOSIM.
}
\references{
Symonds et al. (2016) Journ Applied Microbio  121, p. 1469-1481.
}
\keyword{dataset}
