% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NADA_icmle.R
\name{MLEintvl}
\alias{MLEintvl}
\title{Map maximal antichains to real-valued intervals}
\usage{
MLEintvl(intvls, ml = Maclist(intvls))
}
\arguments{
\item{intvls}{Matrix of intervals (n x 2).}

\item{ml}{Maximal antichain list from \code{\link[=Maclist]{Maclist()}}. If missing, it will be computed.}
}
\value{
A matrix with mapped interval representations for each antichain.
}
\description{
Map maximal antichains to real-valued intervals
}
\keyword{internal}
