% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfit.R
\name{NADA2_survmean}
\alias{NADA2_survmean}
\title{Summary statistics of survival curve (from \code{survival:::survmean})}
\usage{
NADA2_survmean(x, scale = 1, rmean)
}
\arguments{
\item{x}{the result of a call to the survfit function.}

\item{scale}{a numeric value to rescale the survival time, e.g., if the input data to survfit were in days, scale=365 would scale the printout to years.}

\item{rmean}{restricted mean}
}
\description{
computes mean survival times consisent with \code{survival:::survmean}
}
\keyword{internal}
