% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NADA_All.R
\name{NADAList}
\alias{NADAList}
\title{NADA List Class}
\usage{
NADAList(...)
}
\arguments{
\item{...}{Objects to include in the list.}
}
\value{
An object of class \code{NADAList}, which is a list of the input components.
}
\description{
A simple S3 class used to group related NADA objects as a list.
}
