% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NADA_All.R
\docType{data}
\name{NADAprobs}
\alias{NADAprobs}
\title{Default Probability Values for Quantile Estimation}
\format{
Numeric vector
}
\usage{
NADAprobs
}
\description{
A numeric vector of default probabilities used in quantile calculations,
commonly used in censored data methods.
}
\examples{
\dontrun{
NADAprobs
quantile(1:10, probs = NADAprobs)
}
}
\keyword{data}
