% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordranks.R
\name{ordranks}
\alias{ordranks}
\title{Computes ranks of data with one or more detection limits}
\usage{
ordranks(dat.frame, paired = TRUE)
}
\arguments{
\item{dat.frame}{A data frame. Default format is paired = \code{TRUE}, where for 3 chemical parameters the input format is C1 I1 C2 I2 C3 I3, a concentration column followed by its corresponding indicator column.}

\item{paired}{An option to specify paired = \code{FALSE}, where the input format would be C1 C2 C3 I1 I2 I3 where the C columns contain concentrations or a detection limit, and the I columns are their associated indicators, in the same order as the concentration columns.}
}
\value{
Returns a rank for each chemical parameter.
}
\description{
Computes ranks of data (within columns) with one or more DLs, re-censoring at the highest DL if multiple DLs in column are present.
}
\examples{
library(NADA) #For example data
data(PbHeron)

ordranks(PbHeron[,4:15])

}
