% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PbHeron.R
\docType{data}
\name{PbHeron}
\alias{PbHeron}
\title{PbHeron}
\format{
An object of class \code{data.frame} with 27 rows and 15 columns.
}
\source{
Golden et al., 2003, Environmental Toxicology and Chemistry 22, pp. 1517-1524.
}
\usage{
data(PbHeron)
}
\description{
From the \code{NADA} R-Package.

Lead concentrations in the blood and several organs of herons in Virginia.

Objective is to determine the relationships between lead concentrations in the blood and various organs. Do concentrations reflect environmental lead concentrations, as represented by dosing groups? There is one detection limit, at 0.02 ug/g. Used in Chapters 10 and 11 of the NADA book.
}
\references{
Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1st ed. John Wiley and Sons, USA, N.J.
}
\keyword{dataset}
