% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NADA_All.R
\name{pctCen}
\alias{pctCen}
\title{Internal: Calculate Percent Censored Observations}
\usage{
pctCen(obs, censored, na.action = getOption("na.action"))
}
\arguments{
\item{obs}{Numeric observations.}

\item{censored}{Logical vector indicating censored values.}

\item{na.action}{Function or character string specifying NA handling (default getOption("na.action")).}
}
\value{
Percent (0-100) of censored observations.
}
\description{
Internal: Calculate Percent Censored Observations
}
\keyword{internal}
