% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcdf.R
\name{plotcdf}
\alias{plotcdf}
\title{Censored Empirical Cumulative Distribution Function}
\usage{
plotcdf(y.var, cen.var, group = NULL, logscale = FALSE, Ylab = varname)
}
\arguments{
\item{y.var}{The column of data values plus detection limits}

\item{cen.var}{The column of indicators, where 1 (or \code{TRUE}) indicates a detection limit in the \code{y.var} column, and 0 (or \code{FALSE}) indicates a detected value in \code{y.var}.}

\item{group}{Optional - grouping or factor variable. Can be either a text or numeric value indicating the group assignment.}

\item{logscale}{Indicator of whether to plot the data in original y units, or on a log scale.  The default is to use the original units, which usually makes more sense for a cdf plot.}

\item{Ylab}{Optional – input text in quotes to be used as the variable name on the cdf plot.  The default is the name of the \code{y.var} input variable.}
}
\value{
Plots an empirical cumulative distribution function , if \code{group=NULL} then a ECDF with 95\% confidence interval is produced. If \code{group} is identified then ECDFs are produced for each group.
}
\description{
Plots cdfs of one or more groups of censored data.  Illustrates the differences between groups for group tests such as those done using cen1way or cenanova.
}
\examples{
#' library(NADA) #For example data
data(Golden)

# with groups
with(Golden,plotcdf(Liver,LiverCen,DosageGroup))

# all data
with(Golden,plotcdf(Liver,LiverCen))
}
\keyword{CDF}
