% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pollen_Thia.R
\docType{data}
\name{Pollen_Thia}
\alias{Pollen_Thia}
\title{Thiamethoxam concentrations in pollen}
\format{
An object of class \code{data.frame} with 204 rows and 4 columns.
}
\source{
Ontario Ministry of Agriculture, Food and Rural Affairs (Pollen Monitoring Network)
}
\usage{
data(Pollen_Thia)
}
\description{
Thiamethoxam concentrations in pollen from the Ontario Pollen Monitoring Network.
}
\seealso{
\href{https://data.ontario.ca/en/dataset/pollen-monitoring-network-study}{Pollen Monitoring Network Study}

Variables are:
\describe{
\item{Thiamethoxam: }{Thiamethoxam concentration in Concentrations in microgram per gram.}
\item{ThiaCens: }{Censoring indicator.  1 denotes that the value in column 1 is a reporting limit not a specific concentration.}
\item{SamplingEvent: }{A grouping variable from the sample design.  A concentration is from 1 of 4 events in time.}
\item{ThiaAbvBelow: }{A binary variable denoting whether the Thiamethoxam concentration is above or below 0.05 ug/g.}
}
}
\keyword{dataset}
