% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NADA_All.R
\name{print.NADAList}
\alias{print.NADAList}
\title{Show Method for NADAList Objects}
\usage{
\method{print}{NADAList}(x, ...)
}
\arguments{
\item{x}{A \code{NADAList} object.}

\item{...}{further arguments passed to or from other methods. (not used)}
}
\description{
Prints named elements of a \code{NADAList} object.
}
