% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfit.R
\name{quantile_CI}
\alias{quantile_CI}
\title{quantile confidence interval}
\usage{
quantile_CI(n, q, alpha = 0.05)
}
\arguments{
\item{n}{length}

\item{q}{quantile}

\item{alpha}{CI value}
}
\value{
Returns the order statistics and the actual coverage.
}
\description{
quantile confidence interval
}
\keyword{internal}
