% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconLogistic.R
\docType{data}
\name{ReconLogistic}
\alias{ReconLogistic}
\title{ReconLogistic}
\format{
An object of class \code{data.frame} with 423 rows and 8 columns.
}
\source{
Mueller et al., 1997, Journal of Environmental Quality 26, pp. 1223-1230.
}
\usage{
data(ReconLogistic)
}
\description{
Atrazine exceedances (Y/N) above 1 ug/L in streams throughout the Midwestern United States.
Original concentration data were censored to show either greater than (1) or less than (0) 1 ug/L.

Objective -- determine if the pattern of atrazine concentrations above versus below 1 ug/L were related
to seven environmental variables.
}
\references{
Chapter 12 of Helsel, Dennis R. (2012).  Statistics for Censored Environmental Data
using Minitab and R.  John Wiley and Sons, USA, NJ..
}
\keyword{dataset}
