% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NADA_icmle.R
\name{rescaleP}
\alias{rescaleP}
\title{Rescale probability vector}
\usage{
rescaleP(pvec, tiny)
}
\arguments{
\item{pvec}{A probability vector.}

\item{tiny}{Threshold below which values are set to zero.}
}
\value{
A rescaled probability vector.
}
\description{
Rescale probability vector
}
