% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NADA_ros.R
\docType{class}
\name{cenros}
\alias{cenros}
\alias{ros-class}
\title{Class \code{"ros"}}
\format{
An object of class \code{"ros"} with elements:
\describe{
\item{\code{obs}}{The original observations.}
\item{\code{censored}}{Logical vector indicating which observations were censored.}
\item{\code{modeled}}{Vector with uncensored and imputed values.}
\item{\code{pp}}{Plotting positions.}
\item{\code{model}}{The fitted linear model object.}
\item{\code{forwardT}}{Transformation function used.}
\item{\code{reverseT}}{Back-transformation function used.}
}
}
\usage{
cenros(x, ...)
}
\description{
An S3 class returned by the \code{ros()} function, representing the result of regression on order statistics for left-censored data.
}
\section{Functions}{
\itemize{
\item \code{cenros()}: A wrapper for \code{ros()} with simplified argument handling.

}}
\keyword{internal}
