% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSci.R
\name{ROSci}
\alias{ROSci}
\title{Computes confidence intervals on regression on order statistics (ROS) mean}
\usage{
ROSci(cenros.out, conf = 0.95)
}
\arguments{
\item{cenros.out}{an ROS model output object (see details)}

\item{conf}{Confidence coefficient of the interval (Default is 95\%)}
}
\value{
Prints a lower (LCL) and upper (UCL) confidence interval based on the \code{conf} provided (Default is 95\%)
}
\description{
Regression on Ordered Statistics (ROS) is designed to evaluate censored analytical chemistry data. This function evaluates ROS results and computes the confidence interval around the mean.
}
\details{
This function uses an ROS model output based on the \code{ros} funtion in the \code{NADA} package.

For more detail on ROS modeling see the \code{ros} help file (\code{?NADA::ros}).

For implementation of \code{ROSci(...)} see the examples below.
}
\examples{
library(NADA) #For example data

data(HgFish)
myros <- NADA::ros(HgFish$Hg,HgFish$HgCen)

summary(myros)

# ROS Mean
mean(myros)

# 95\% CI around the ROS mean
ROSci(myros)
}
\references{
Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1st ed. John Wiley and Sons, USA, N.J.

Lee, L., Helsel, D., 2005. Statistical analysis of water-quality data containing multiple detection limits: S-language software for regression on order statistics. Computers & Geosciences 31, 1241–1248. \url{https://doi.org/10.1016/j.cageo.2005.03.012}

Zhou, X.-H., Gao, S., 1997. Confidence Intervals for the Log-Normal Mean. Statistics in Medicine 16, 783–790. <https://doi.org/10.1002/(SICI)1097-0258(19970415)16:7<783::AID-SIM488>3.0.CO;2-2>
}
\seealso{
\link[NADA:ros]{NADA::ros}
}
