% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShePyrene.R
\docType{data}
\name{ShePyrene}
\alias{ShePyrene}
\title{ShePyrene}
\format{
An object of class \code{data.frame} with 56 rows and 2 columns.
}
\source{
She, N., 1997, Analyzing censored water quality data using a nonparametric approach.  Journal of the American Water Resources Association, 33, pp615--624.
}
\usage{
data(ShePyrene)
}
\description{
Originally in the \code{NADA} R-Package.

Pyrene concentrations in milligrams per liter from 20 water-quality monitoring stations in the Puget Sound of Washington State, USA.

Used for characterizing priority pollutant concentrations in sediments of Puget Sound by computing summary statisitics.  Contains eight detection limits with 11 nondetects out of 56 total measurements.
}
\references{
Helsel, D.R., 2005.Nondectects and Data Analysis; Statistics for censored environmental data. John Wiley & Sons, USA, N.J.
}
\keyword{dataset}
