% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCE2.R
\docType{data}
\name{TCE2}
\alias{TCE2}
\title{TCE2}
\format{
An object of class \code{data.frame} with 222 rows and 6 columns.
}
\source{
Eckhardt et al., 1989,	USGS Water Resources Investigations Report 86-4142.
}
\usage{
data(TCE2)
}
\description{
TCE concentrations (ug/L) in ground waters of Long Island, New York.
Categorized by two of the three land use types (medium and high density residential) surrounding the wells and measured in the study.

Objective -- determine if TCE concentrations differ in ground water under two land use types.
There are four detection limits, at 1,2,4 and 5 ug/L.
}
\references{
The full dataset with three land use types is used in Chapter 10 of Statistics for
Censored Environmental Data using Minitab and R, by D.R.Helsel; Wiley (2012).
}
\keyword{dataset}
