% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCEReg.R
\docType{data}
\name{TCEReg}
\alias{TCEReg}
\title{TCE Ground Waters of Long Island — with Explanatory Variables}
\format{
A data frame with variables related to TCE groundwater concentrations.
}
\source{
Eckhardt et al., 1989. USGS Water Resources Investigations Report 86-4142.
}
\usage{
data(TCEReg)
}
\description{
TCE concentrations (µg/L) in ground waters of Long Island, New York, along with several possible explanatory variables.
}
\details{
The objective is to determine if concentrations are related to one or more explanatory variables.

There are four detection limits at 1, 2, 4, and 5 µg/L. One column indicates whether concentrations are above or below 5.
Used in Chapter 12 of the NADA book.
}
\references{
Helsel, Dennis R. (2005). \emph{Nondetects and Data Analysis: Statistics for Censored Environmental Data}. John Wiley and Sons, USA, NJ.
}
\keyword{datasets}
