% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usc.R
\name{Usc}
\alias{Usc}
\title{Interval-censored U-Score}
\usage{
Usc(ylo, yhi, rnk = T)
}
\arguments{
\item{ylo}{the lower end of the concentration interval}

\item{yhi}{the upper end of the concentration interval}

\item{rnk}{A \code{TRUE}/\code{FALSE} variable on whether to compute the multivariate pattern on the uscores, or the ranks of the uscores.  Default is rnk=\code{TRUE}, use the ranks. rnk = \code{FALSE} returns the uscores.}
}
\description{
Interval-censored computation of uscores and their ranks for 1 parameter.
}
