% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscores.R
\name{uscores}
\alias{uscores}
\title{U-score (multiple values)}
\usage{
uscores(dat.frame, paired = TRUE, rnk = TRUE)
}
\arguments{
\item{dat.frame}{A data frame. Default format is paired = \code{TRUE}, where for 3 chemical parameters the input format is C1 I1 C2 I2 C3 I3, a concentration column followed by its corresponding indicator column.}

\item{paired}{When paired = \code{FALSE}, the input format is C1 C2 C3 I1 I2 I3 where the C columns contain concentrations or a detection limit, and the I columns are their associated indicators, in the same order as the concentration columns.}

\item{rnk}{A logical \code{TRUE}/\code{FALSE} variable on whether to compute the multivariate pattern on the uscores, or the ranks of the uscores.  Default is rnk=\code{TRUE}, use the ranks. rnk = \code{FALSE} returns the uscores.}
}
\value{
U-score for each chemical parameter
}
\description{
Computes uscores of data (requires 2 columns) with one or more DLs
}
\examples{
library(NADA) #For example data
data(Golden)

uscores(Golden[,4:15])
}
