% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anosimPlot.R
\name{anosimPlot}
\alias{anosimPlot}
\title{Permutation analysis of similarities between groups}
\usage{
anosimPlot(
  ano.out,
  hcol = "light blue",
  title = "Histogram of anosim permutations"
)
}
\arguments{
\item{ano.out}{an \code{anosim} object. See details and example}

\item{hcol}{color of histogram}

\item{title}{title of histogram}
}
\value{
prints a histogram of permutation tests
}
\description{
Permutation analysis of similarities between groups
}
\examples{
data(PbHeron)

# ROS model for each group
PbHeron.high <- with(subset(PbHeron,DosageGroup=="High"),NADA::ros(Blood,BloodCen))
PbHeron.high <- data.frame(PbHeron.high)
PbHeron.high$DosageGroup <- "High"

PbHeron.low <- with(subset(PbHeron,DosageGroup=="Low"),NADA::ros(Blood,BloodCen))
PbHeron.low <- data.frame(PbHeron.low)
PbHeron.low$DosageGroup <- "Low"

PbHeron.ros=rbind(PbHeron.high,PbHeron.low)

# ANOSIM analysis
library(vegan)
PbHeron.anosim <- with(PbHeron.ros,anosim(modeled,DosageGroup))
summary(PbHeron.anosim)

# Plot
anosimPlot(PbHeron.anosim)
}
\references{
Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1st ed. John Wiley and Sons, USA, N.J.

Oksanen, J., Guillaume, F., 2018. Vegan: ecological diversity. CRAN R-Project.\url{https://cran.r-project.org/web/packages/vegan/index.html}
}
\seealso{
\link[vegan:anosim]{vegan::anosim}
}
