% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATS.R
\name{ATS}
\alias{ATS}
\title{Akritas-Theil-Sen line for censored data.}
\usage{
ATS(
  y.var,
  y.cen,
  x.var,
  x.cen = rep(0, times = length(x.var)),
  LOG = TRUE,
  retrans = FALSE,
  xlabel = NULL,
  ylabel = NULL
)
}
\arguments{
\item{y.var}{The column of y (response variable) values plus detection limits}

\item{y.cen}{The y-variable indicators, where 1 (or `TRUE`) indicates a detection limit in the `y.var` column, and `0` (or `FALSE`) indicates a detected value in `y.var`.}

\item{x.var}{The column of x (explanatory variable) values plus detection limits}

\item{x.cen}{The x-variable indicators, where 1 (or `TRUE`) indicates a detection limit in the `x.var` column, and `0` (or `FALSE`) indicates a detected value in `x.var`.}

\item{LOG}{Indicator of whether to compute the ATS line in the original y units, or for their logarithms.  The default is to use the logarithms (`LOG = TRUE`).  To compute in original units, specify the option `LOG = FALSE` (or `LOG = 0`).}

\item{retrans}{Indicator of whether to retransform the plot and line back to original Y-variable units.  Not needed when `LOG = FALSE`. `retrans = FALSE` & `LOG = TRUE` draws the plot in logY units. `retrans =  TRUE` & `LOG = TRUE` draws the plot in original Y units.}

\item{xlabel}{Custom label for the x axis of plots.  Default is x variable column name.}

\item{ylabel}{Custom label for the y axis of plots.  Default is y variable column name.}
}
\value{
Draws a scatterplot with the fitted line superimposed.
}
\description{
Computes Kendall's tau and the Akritas-Theil-Sen (ATS) line for censored data.  For one x variable regression.
}
\examples{
# Both y and x are censored
data(Golden)
with(Golden, ATS(Blood, BloodCen, Kidney, KidneyCen))

# x is not censored
data(TCEReg)
with(TCEReg, ATS(TCEConc, TCECen, PctIndLU))
}
\keyword{censored}
\keyword{trend}
