% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryDiss.R
\name{binaryDiss}
\alias{binaryDiss}
\title{Binary dissimilarity coefficient}
\usage{
binaryDiss(dat.frame)
}
\arguments{
\item{dat.frame}{A data frame containing only the 0/1 columns.}
}
\value{
Returns a binary dissimilarity matrix.
}
\description{
Computes a simple matching dissimilarity coefficient
}
\examples{
data(PbHeron)

binaryDiss(PbHeron$LiverCen)
}
\seealso{
\link[vegan:designdist]{vegan::designdist}
}
