% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryMDS.R
\name{binaryMDS}
\alias{binaryMDS}
\title{Plot Nonmetric Multidimensional Scaling of censored data}
\usage{
binaryMDS(dat.frame, group = NULL, title = NULL, legend.pos = "bottomleft")
}
\arguments{
\item{dat.frame}{A data frame containing only the 0/1 columns.}

\item{group}{Optional grouping variable. Sites will be represented by different colored symbols for each group.}

\item{title}{Optional title for the NMDS graph.}

\item{legend.pos}{For when group is specified, the location of the legend on the graph showing the colors representing each group’s data.  Default is “bottomleft”.  Alternatives are “topright” and “centerleft”, etc.}
}
\value{
Print NMDS plot of censored data groupings.
}
\description{
Plots an NMDS of a matrix of 0s and 1s, the censoring indicator columns for multiple variables.  Use the highest censoring limit within each column.  May have different censoring levels in different columns.
}
\examples{
library(NADA) #For example data
data(Golden)

# without group specified
binaryMDS(Golden[,4:15])

#' # With Group argument
binaryMDS(Golden[,4:15],group=Golden$DosageGroup)
}
\seealso{
\link[vegan:metaMDS]{vegan::metaMDS}
}
