% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarySim.R
\name{binarySim}
\alias{binarySim}
\title{Binary similarity coefficient matrix}
\usage{
binarySim(dat.frame)
}
\arguments{
\item{dat.frame}{A data frame containing only the 0/1 columns.}
}
\value{
Returns a binary similarity matrix.
}
\description{
Computes a simple matching similarity coefficient
}
\examples{
data(PbHeron)

binarySim(PbHeron$LiverCen)

}
\references{
Helsel, D.R., 2011. Statistics for Censored Environmental Data using Minitab and R, 2nd ed. John Wiley & Sons, USA, N.J.
}
\seealso{
\link[vegan:designdist]{vegan::designdist}
}
