% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cencorreg.R
\name{cencorreg}
\alias{cencorreg}
\title{Correlation and Regression with censored data}
\usage{
cencorreg(y.var, cen.var, x.vars, LOG = TRUE, verbose = 2)
}
\arguments{
\item{y.var}{The column of y (response variable) values plus detection limits.}

\item{cen.var}{The column of indicators, where 1 (or \code{TRUE}) indicates a detection limit in the \code{y.var} column, and 0 (or \code{FALSE}) indicates a detected value is in \code{y.var}.}

\item{x.vars}{One or more uncensored explanatory variable(s). See Details}

\item{LOG}{Indicator of whether to compute the regression in the original y units, or on their logarithms.  The default is to use the logarithms (\code{LOG = TRUE}).  To compute in original units, specify the option \code{LOG = FALSE} (or \code{LOG = 0}).}

\item{verbose}{default \code{verbose=2}, see details.}
}
\value{
When \code{x.vars} is one variable, likelihood, rescaled likelihood and McFaddens correlation coefficient (\code{R}) are printed.
When \code{x.vars} is a \code{data.frame} of more than one variable, likelihood, rescaled likelihood and McFaddens coefficent of determination (\code{R2}) are printed.

Model coefficients (intercept and slopes), Chi-Squared statistic and p-value for the test that all slope coefficients equal zero (overall test), and model AIC and BIC are provided.

A Q-Q plot of model residuals with corresponding Shapiro-Francia W and p-value are plotted for evaluation of model distributional assumptions when \code{verbose=2} (the default).
}
\description{
Computes three parametric correlation coefficients for one X variable and the corresponding R squared for multiple X variables, and a regression equation for censored data.
}
\details{
\code{x.vars}: If one x variable only, enter its name.  If multiple x variables, enter the name of a data frame of columns of the x variables. Only columns used as \code{X} variables in the regression are allowed. Create this by \code{x.frame <- data.frame (Temp, Flow, Time)} for 3 variables (temperature, flow and time) used as the \code{X} variables in the regression.

AIC and BIC are printed to help evaluate the ‘best’ regression model.

The default is that the Y variable will be log transformed.

\code{verbose} option. Default is 2 which provides full output in the console and qqplots in a graphics window. A value of 1 only provides partial results in the console and no qqplots. A value of 0 provides no output; the returning computations will be stored in the specified object.
}
\examples{

data(Brumbaugh)

# One variable
cencorreg(Brumbaugh$Hg,Brumbaugh$HgCen,Brumbaugh$SedMeHg)

# More than one variable for demonstration purposes
cencorreg(Brumbaugh$Hg,Brumbaugh$HgCen,Brumbaugh[,c("SedMeHg","PctWetland")])
}
\references{
Helsel, D.R., 2011. Statistics for censored environmental data using Minitab and R, 2nd ed. John Wiley & Sons, USA, N.J.

Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1st ed. John Wiley and Sons, USA, N.J.
}
\seealso{
\link[survival:survreg]{survival::survreg}
}
