% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrend.R
\name{centrend}
\alias{centrend}
\title{Trend analysis of residuals of censored data}
\usage{
centrend(y.var, y.cens, x.var, time.var, link = "identity", Smooth = "cs")
}
\arguments{
\item{y.var}{The column of y (response variable) values plus detection limits}

\item{y.cens}{The column of indicators, where 1 (or \code{TRUE}) indicates a detectionlimit in the \code{y.var} column, and 0 (or \code{FALSE}) indicates a detected value in \code{y.var}.}

\item{x.var}{Column of a covariate (not time).  \code{y.var} will be smoothed versus \code{x.var} and residuals taken to subtract out the relationship between \code{y} and \code{x}.}

\item{time.var}{Column of the time variable, either a sequence of days or decimal times, etc.  Will be the scale used for time in ATS trend analysis.}

\item{link}{Default = \verb{“identity”} which means the original data.}

\item{Smooth}{Type of smoother used in the GAM.Default is \verb{“cs”}, shrinkage cubic regression splines.See details}
}
\value{
Prints three plots (Data with GAM Smooth, Residuals from GAM Smooth and ATS trend.)

Returns GAM residuals, and ATS results
}
\description{
Trend analysis of residuals of censored data
}
\details{
Default \code{link} is identity, other options are available see \code{cenGAM::tobit1} for more options.

Default \code{Smooth} is \code{"cs"} for shrinkage cubic regression splines. See \code{mgcv::smooth.terms} for other smoothing terms.
}
\examples{

data(Brumbaugh)

Brumbaugh$time=1:nrow(Brumbaugh)

with(Brumbaugh,centrend(Hg,HgCen,SedTotHg,time.var=time))
}
\references{
Helsel, D.R., 2011. Statistics for censored environmental data using Minitab and R, 2nd ed. John Wiley & Sons, USA, N.J.

Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1st ed. John Wiley and Sons, USA, N.J.
}
\seealso{
\link[mgcv:gam]{mgcv::gam}
}
\keyword{GAM}
\keyword{analysis}
\keyword{spline}
\keyword{trend}
