% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfit.R
\name{cfit}
\alias{cfit}
\title{Compute an ECDF for Censored Data}
\usage{
cfit(
  y1,
  y2,
  conf = 0.95,
  qtls = c(0.1, 0.25, 0.5, 0.75, 0.9),
  Cdf = TRUE,
  printstats = TRUE,
  Ylab = NULL
)
}
\arguments{
\item{y1}{lowest possible concentration (interval-censored) or concs+DLs for indicator format}

\item{y2}{highest possible concentration (interval-censored) or censoring indicator (logical. 1 or \code{TRUE} = censored, 0 or FALSE = detected) for indicator format.}

\item{conf}{the confidence coefficient for confidence intervals around the Kaplan-Meier mean and median.}

\item{qtls}{Probabilities for the quantiles to be estimated.  The defaults are listed above.  You may add and/or substitute probabilities.}

\item{Cdf}{Logical \code{TRUE}/\code{FALSE} Indicator of whether to plot the data a cumulative distribution function (cdf) using Kaplan-Meier quantiles.}

\item{printstats}{Logical \code{TRUE}/\code{FALSE} Option of whether to print the resulting statisics in the console window, or not}

\item{Ylab}{Optional – input text in quotes to be used as the variable name on the cdf plot.  The default is the name of the \code{y1} input variable.}
}
\value{
If \code{printstats=TRUE} Based on the provided \code{conf} value, Kaplan-Meier summary statistics (\code{mean},\code{sd},\code{median}), lower and upper confidence intervals around the mean and median value, sample size and percent of censored samples. Additionally specified quartile values are returned.

If \code{Cdf=TRUE} eCDF plot of censored data is printed.
}
\description{
Computes an estimate of an empirical cumulative distribution function (ECDF) for censored data
}
\examples{
library(NADA) #For example data

data(Cadmium)

cfit(Cadmium$Cd,Cadmium$CdCen)
}
\references{
Helsel, D.R., 2005. Nondetects and Data Analysis: Statistics for Censored Environmental Data, 1st ed. John Wiley and Sons, USA, N.J.
}
\seealso{
\link[survival:survfit]{survival::survfit} \link[NADA:cenfit]{NADA::cenfit} \link[NADA:cendiff]{NADA::cendiff}
}
